Move back Poly, _parse_shape_spec into masking.py to simplify diff
