+++ jax/lax/lax.py
@@ -40,7 +40,8 @@ from ..config import flags
 from ..core import Primitive
 from ..abstract_arrays import (UnshapedArray, ShapedArray, ConcreteArray,
                                AbstractToken, array_types, make_shaped_array,
-                               raise_to_shaped, abstract_token, to_index)
+                               raise_to_shaped, abstract_token)
+from ..interpreters.masking import to_index
 from ..interpreters import partial_eval as pe
 from ..interpreters import xla
 from ..interpreters import pxla
