+++ tests/masking_test.py
@@ -21,9 +21,8 @@ from unittest import SkipTest
 
 import numpy as onp
 from absl.testing import absltest
-from jax.interpreters.masking import shape_as_value
-from jax import numpy as np, test_util as jtu, mask, vmap, jit, grad, lax, \
-  ShapeError
+from jax.interpreters.masking import shape_as_value, ShapeError
+from jax import numpy as np, test_util as jtu, mask, vmap, jit, grad, lax
 from jax.config import config
 config.parse_flags_with_absl()
